+++ Kitodo/src/main/java/org/kitodo/production/services/data/ProcessService.java
@@ -220,6 +220,21 @@ public class ProcessService extends ProjectSearchService<Process, ProcessDTO, Pr
         RULESET_CACHE_FOR_CREATE_FROM_CALENDAR.clear();
     }
 
+    /**
+     * Checks if an imported Process should be created with Tasks and removes them if not,
+     * depending on the configuration of the doctype.
+     * @param process the process to check.
+     * @param docType the doctype to check in the ruleset.
+     */
+    public static void checkTasks(Process process, String docType) throws IOException {
+        // remove tasks from process, if doctype is configured not to use a workflow
+        Collection<String> divisionsWithNoWorkflow = ServiceManager.getRulesetService()
+                .openRuleset(process.getRuleset()).getDivisionsWithNoWorkflow();
+        if (divisionsWithNoWorkflow.contains(docType)) {
+            process.getTasks().clear();
+        }
+    }
+
     @Override
     public Long countDatabaseRows() throws DAOException {
         return countDatabaseRows("SELECT COUNT(*) FROM Process");
